Two fixes/updates
First issue appears if you addData twice with the firstLineTitles = false. On the second load the titles text would be overwritten with prop-0, prop-1, ... This was due to the this._propertiesNames.indexOf(prop) >= 0 which should not be set.

Second the base method for addData returned the added layer, I use this and had to add the return (which should have been there anyway).